+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
+## [1.2.0] - 2023-01-06
+
+### Added
+
+- Changed a `SocketMessageRouter`'s method from `hasExists` to `hasAction`.
+- Added method `remove` to `SocketMessageRouter` to allow a graceful removal of action handler.
+
 ## [1.1.8] - 2022-09-02
 
 ### Fixed